Share the lighting panel; light the CRT scene properly
The stress tests' lighting controls (environment menu, skybox toggle,
sky blur, exposure, IBL intensity, environment rotation) move into a
shared, self-contained LightingPanel that applies directly to the
scene it is given; the stress tests drop their local copy and the
widget-texture example gains the panel with the skybox enabled, so the
tube's glare can be tested under any environment. The example floor was
a rough metal (metallicFactor defaults to 1.0), reading as a mirror; it
is now a dielectric at 0.3 roughness.
